Add an error message for low coin count
In this commit, an alert feature has been added to the application that handles coin management. Now, the application checks if the count of 20Kč coins is less than 15 and if it is, a warning message is displayed urging the user to replenish them. The HTML template has also been updated to incorporate this new feature.